+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
----
-
 ### Added
 
 ### Changed
+
+- Enhance the MQTT client with subscription handling, HA birth topic support, and connection stability improvements
 
 ### Removed